requirements.txt
streamlit 
st-gsheets-connection
pillow